Remove check for su using a hardcoded path.
Upstream project does a check for /system/bin/su and displays a warning.  On CM7 su is loced at /system/xbin/su so this will fire a warning every time the app starts.

This commit removes that check.  The real fix will be to try to execute su and see if that fails and issue a warning like SetCPU or TitaniumBackup do at start, but this will work for now